+++ Blink.py
@@ -1,0 +1,122 @@
+from scipy.spatial import distance as dist
+from imutils.video import FileVideoStream
+from imutils.video import VideoStream
+from imutils import face_utils
+import numpy as np
+import imutils
+import time
+import dlib
+import cv2
+
+class Blink:
+
+    def __init__(self, detector, predictor, 
+                ear_thresh = 0.2, ear_consec_frame = 2,ear= None, blinks = None, source=None ):
+        self.source = source
+        self.detector = detector
+        self.predictor = predictor
+        self.ear_thresh = ear_thresh
+        self.ear_consec_frame = ear_consec_frame
+        self.ear = ear
+        self.blinks = blinks
+
+    def detect(self):
+        # grab the indexes of the facial landmarks for the left and
+        # right eye, respectively
+        (lStart, lEnd) = face_utils.FACIAL_LANDMARKS_IDXS["left_eye"]
+        (rStart, rEnd) = face_utils.FACIAL_LANDMARKS_IDXS["right_eye"]
+
+        self.source = VideoStream(src=0).start()
+        cont_frames = 0
+        self.blinks = 0
+
+        # loop over frames from the video stream
+        while True:
+            # grab the frame from the threaded video file stream, resize
+            # it, and convert it to grayscale
+            # channels)
+            frame = self.source.read()
+            frame = imutils.resize(frame, width=800)
+            gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
+
+            # detect faces in the grayscale frame
+            rects = self.detector(gray, 0)
+
+            # loop over the face detections
+            for rect in rects:
+                # determine the facial landmarks for the face region, then
+                # convert the facial landmark (x, y)-coordinates to a NumPy
+                # array
+                shape = self.predictor(gray, rect)
+                shape = face_utils.shape_to_np(shape)
+
+                # extract the left and right eye coordinates, then use the
+                # coordinates to compute the eye aspect ratio for both eyes
+                leftEye = shape[lStart:lEnd]
+                rightEye = shape[rStart:rEnd]
+                leftEAR = self.eye_aspect_ratio(leftEye)
+                rightEAR = self.eye_aspect_ratio(rightEye)
+
+                # average the eye aspect ratio together for both eyes
+                self.ear = (leftEAR + rightEAR) / 2.0
+
+                # compute the convex hull for the left and right eye, then
+                # visualize each of the eyes
+                leftEyeHull = cv2.convexHull(leftEye)
+                rightEyeHull = cv2.convexHull(rightEye)
+                cv2.drawContours(frame, [leftEyeHull], -1, (0, 255, 0), 1)
+                cv2.drawContours(frame, [rightEyeHull], -1, (0, 255, 0), 1)
+
+                # check to see if the eye aspect ratio is below the blink
+                # threshold, and if so, increment the blink frame counter
+                if self.ear < self.ear_thresh:
+                    cont_frames += 1
+
+                # otherwise, the eye aspect ratio is not below the blink
+                # threshold
+                else:
+                    # if the eyes were closed for a sufficient number of
+                    # then increment the total number of blinks
+                    if cont_frames >= self.ear_consec_frame:
+                        self.blinks += 1
+
+                    # reset the eye frame counter
+                    cont_frames = 0
+
+                # draw the total number of blinks on the frame along with
+                # the computed eye aspect ratio for the frame
+                cv2.putText(frame, "Blinks: {}".format(self.blinks), (10, 30),
+                    cv2.FONT_HERSHEY_SIMPLEX, 0.7, (0, 0, 255), 2)
+                cv2.putText(frame, "EAR: {:.2f}".format(self.ear), (300, 30),
+                    cv2.FONT_HERSHEY_SIMPLEX, 0.7, (0, 0, 255), 2)
+                (x, y, w, h) = face_utils.rect_to_bb(rect)
+                face = frame[y : y + h, x : x + w]
+                face = imutils.resize(face, width = 400)
+                cv2.imshow("Face", face)
+
+            # show the frame
+            cv2.imshow("Frame", frame)
+            key = cv2.waitKey(1) & 0xFF
+        
+            # if the `q` key was pressed, break from the loop
+            if key == ord("q"):
+                break
+
+
+
+
+    def eye_aspect_ratio(self,eye):
+        # compute the euclidean distances between the two sets of
+        # vertical eye landmarks (x, y)-coordinates
+        A = dist.euclidean(eye[1], eye[5])
+        B = dist.euclidean(eye[2], eye[4])
+
+        # compute the euclidean distance between the horizontal
+        # eye landmark (x, y)-coordinates
+        C = dist.euclidean(eye[0], eye[3])
+
+        # compute the eye aspect ratio
+        ear = (A + B) / (2.0 * C)
+
+        # return the eye aspect ratio
+        return ear
