--- logfile.txt
@@ -1,24 +1,0 @@
-   Date        Time     Command
-   
-2019-09-29    0:14:35    palm
-2019-09-29    0:14:38    palm
-2019-09-29    0:14:41    palm
-2019-09-29    0:14:47    palm
-2019-09-29    0:14:50    palm
-2019-09-29    0:14:53    palm
-2019-09-29    0:14:56    palm
-2019-09-29    0:15:05    palm
-2019-09-29    0:15:08    palm
-2019-09-29    0:15:14    palm
-2019-09-29    0:19:10    palm
-2019-09-29    0:19:13    palm
-2019-09-29    0:19:16    palm
-2019-09-29    0:19:19    palm
-2019-09-29    0:20:21    palm
-2019-09-29    0:20:24    palm
-2019-09-29    0:20:27    palm
-2019-09-29    0:20:30    palm
-2019-09-29    0:20:33    palm
-2019-09-29    0:20:36    palm
-2019-09-29    0:20:39    palm
-2019-09-29    0:20:58    palm
